+++ packages/chakra/dev/spel2js.d.ts
@@ -1,0 +1,3 @@
+declare module 'spel2js' {
+  export { SpelExpressionEvaluator } from 'spel2js';
+}