+++ packages/dnd/src/hooks/useRuleDnD.ts
@@ -35,6 +35,7 @@ export const useRuleDnD = (params: UseRuleDndParams): UseRuleDnD => {
     useDrag,
   });
 
+  // eslint-disable-next-line react-compiler/react-compiler
   const [{ isOver, dropMonitorId, dropEffect }, drop] = useDrop!<
     DraggedItem,
     DropResult,