+++ packages/react-querybuilder/src/hooks/usePrevious.ts
@@ -1,3 +1,4 @@
+/* eslint-disable react-compiler/react-compiler */
 import { useRef } from 'react';
 
 /**