+++ packages/react-querybuilder/src/hooks/useValueEditor.test.ts
@@ -1,4 +1,5 @@
 import { renderHook } from '@testing-library/react';
+import type { ParseNumberMethod, ParseNumbersPropConfig } from '../types';
 import { useValueEditor } from './useValueEditor';
 
 it('calls handleOnChange when operator is not "between"/"in" and value is an array', async () => {
@@ -38,6 +39,45 @@ it('sets valueAsArray when operator is "between"', async () => {
     })
   );
   expect(hr.result.current.valueAsArray).toEqual([12, 14]);
+});
+
+describe('parseNumbers', () => {
+  it.each([
+    { pn: true, text: 'strict', number: 'strict' },
+    { pn: false, text: false, number: false },
+    { pn: 'enhanced', text: 'enhanced', number: 'enhanced' },
+    { pn: 'enhanced-limited', text: false, number: 'enhanced' },
+    { pn: 'native', text: 'native', number: 'native' },
+    { pn: 'native-limited', text: false, number: 'native' },
+    { pn: 'strict', text: 'strict', number: 'strict' },
+    { pn: 'strict-limited', text: false, number: 'strict' },
+  ] satisfies {
+    pn: ParseNumbersPropConfig;
+    text: ParseNumberMethod;
+    number: ParseNumberMethod;
+  }[])('processes $pn correctly', async ({ pn, number, text }) => {
+    const handleOnChangeText = jest.fn();
+    const hrText = renderHook(() =>
+      useValueEditor({
+        handleOnChange: handleOnChangeText,
+        operator: '=',
+        inputType: 'text',
+        parseNumbers: pn,
+      })
+    );
+    expect(hrText.result.current.parseNumberMethod).toBe(text);
+
+    const handleOnChangeNumber = jest.fn();
+    const hrNumber = renderHook(() =>
+      useValueEditor({
+        handleOnChange: handleOnChangeNumber,
+        operator: '=',
+        inputType: 'number',
+        parseNumbers: pn,
+      })
+    );
+    expect(hrNumber.result.current.parseNumberMethod).toBe(number);
+  });
 });
 
 it('does not call handleOnChange when type is "multiselect"', async () => {